moo
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7570 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/TextUtils.h
@@ -88,7 +88,7 @@ public:
       // push the  part up to 
       tempStream << in.substr(beginPos,endPos-beginPos); 
       tempStream << withMe;
-      beginPos = endPos + replaceMe.size();
+      beginPos = (long int)(endPos + replaceMe.size());
       endPos = (long int)in.find(replaceMe,beginPos);
     } 
     tempStream << in.substr(beginPos);
